+++ activity/src/main/java/com/kairos/service/pay_out/PayOutService.java
@@ -135,7 +135,7 @@ public class PayOutService {
         shift.setPayoutCtaBonusMinutes(shift.getActivities().stream().mapToInt(shiftActivity -> shiftActivity.getPayoutCtaBonusMinutes()).sum());
         shift.setPayoutCtaBonusMinutes(shift.getPayoutCtaBonusMinutes() + shift.getBreakActivities().stream().mapToInt(shiftActivity -> shiftActivity.getPayoutCtaBonusMinutes()).sum());
         shift.setPayoutPerShiftCTADistributions(ObjectMapperUtils.copyCollectionPropertiesByMapper(shiftWithActivityDTO.getPayoutPerShiftCTADistributions(), PayOutPerShiftCTADistribution.class));
-        int ctaBonusOfShift = shift.getPayoutPerShiftCTADistributions().stream().mapToInt(payOutPerShiftCTADistribution -> payOutPerShiftCTADistribution.getMinutes()).sum();
+        int ctaBonusOfShift = shift.getPayoutPerShiftCTADistributions().stream().mapToInt(payOutPerShiftCTADistribution -> (int)payOutPerShiftCTADistribution.getMinutes()).sum();
         shift.setPayoutCtaBonusMinutes(shift.getPayoutCtaBonusMinutes() + ctaBonusOfShift);
         shift.setPlannedMinutesOfPayout(shift.getPlannedMinutesOfPayout()+ctaBonusOfShift);
     }
