+++ kcommons/src/main/java/com/kairos/dto/activity/time_bank/time_bank_basic/time_bank/CTADistributionDTO.java
@@ -21,6 +21,6 @@ public class CTADistributionDTO  implements Serializable {
     private BigInteger id;
     private String name;
     private double minutes;
-    private float cost;
+    private double cost;
 
 }
